pyproject.toml
[project]
name = "prathamesh-ghatole-github-io"
version = "1.4.0"
description = "A static site for my personal portfolio and blog."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "markdown>=3.10",
    "python-frontmatter>=1.1.0",
]

[dependency-groups]
dev = [
    "livereload>=2.7.0",
]
